requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.0
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycparser==2.22
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.10.0
urllib3==1.26.18
